requirements.txt
tensorflow-gpu==2.2.0
numpy==1.18.5
Pillow==7.1.2
Flask==1.1.2
requests==2.24.0
uWSGI==2.0.19.1